Boite sélection de commune dans la nouvelle zone de filtres
