+++ webapp/components/PollutionMapBase.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import { useEffect, useMemo } from "react";
-import ReactMapGl, { MapLayerMouseEvent } from "react-map-gl/maplibre";
-import maplibregl from "maplibre-gl";
+import { useEffect, useMemo, useRef } from "react";
+import ReactMapGl, { MapLayerMouseEvent, useMap } from "react-map-gl/maplibre";
+import maplibregl, { Feature } from "maplibre-gl";
 import "maplibre-gl/dist/maplibre-gl.css";
 import { Protocol } from "pmtiles";
 import layers from "protomaps-themes-base";
@@ -15,7 +15,7 @@ import {
 type PollutionMapBaseLayerProps = {
   year: string;
   categoryType: string;
-  selectedCommune: any | null;
+  selectedCommune: Feature | null;
   onFeatureClick: (feature: any) => void;
 };
 
@@ -25,6 +25,8 @@ export default function PollutionMapBaseLayer({
   selectedCommune,
   onFeatureClick,
 }: PollutionMapBaseLayerProps) {
+  const mapRef = useRef(null);
+
   useEffect(() => {
     // adds the support for PMTiles
     const protocol = new Protocol();
@@ -42,12 +44,17 @@ export default function PollutionMapBaseLayer({
       onFeatureClick(event.features[0]);
     }
   }
+  const { current: map } = useMap();
 
   useEffect(() => {
     if (selectedCommune) {
-      console.log("Should zoom to:", selectedCommune);
+      console.log("Should zoom to:", selectedCommune.geometry.coordinates);
+      mapRef?.current?.flyTo({
+        center: selectedCommune.geometry.coordinates,
+        zoom: 9,
+      });
     }
-  }, [selectedCommune]);
+  }, [selectedCommune, map]);
 
   const mapStyle = useMemo(() => {
     const dynamicLayers: maplibregl.LayerSpecification[] = [
@@ -70,9 +77,13 @@ export default function PollutionMapBaseLayer({
           "fill-opacity": 0.5,
         },
         // Ajout d'un filtre pour mettre en évidence la commune sélectionnée si présente
-        ...(selectedCommune
+        ...(0 && selectedCommune
           ? {
-              filter: ["==", ["get", "code_insee"], selectedCommune.code_insee],
+              filter: [
+                "==",
+                ["get", "code_insee"],
+                selectedCommune.properties?.id,
+              ],
             }
           : {}),
       },
@@ -92,6 +103,7 @@ export default function PollutionMapBaseLayer({
       mapLib={maplibregl}
       onClick={onClick}
       interactiveLayerIds={["polluants"]}
+      ref={mapRef}
     />
   );
 }
